requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
torch>=2.1.0
